update details page
