+++ src/actions/index.js
@@ -1,12 +1,9 @@
 /* eslint-disable no-console */
-import {
-  getLocation as getLoc,
-  updateLocation as updateLoc,
-  updatePhone as updatePh,
-  createPhone as createPh,
-} from '../services/api';
+import * as api from '../services/api';
 
 export const GET_LOCATION_RESPONSE = 'GET_LOCATION_RESPONSE';
+export const GET_TAXONOMY_RESPONSE = 'GET_TAXONOMY_RESPONSE';
+export const SELECT_CATEGORIES_ACTION = 'SELECT_CATEGORIES_ACTION';
 export const OPTIMISTIC_UPDATE_LOCATION = 'OPTIMISTIC_UPDATE_LOCATION';
 export const ROLLBACK_UPDATE_LOCATION = 'ROLLBACK_UPDATE_LOCATION';
 export const OPTIMISTIC_UPDATE_PHONE = 'OPTIMISTIC_UPDATE_PHONE';
@@ -14,15 +11,34 @@ export const OPTIMISTIC_CREATE_PHONE = 'OPTIMISTIC_CREATE_PHONE';
 export const CREATE_PHONE_SUCCESS = 'CREATE_PHONE_SUCCESS';
 
 export const getLocation = locationId => (dispatch) => {
-  getLoc({
-    id: locationId,
-  })
+  api
+    .getLocation({
+      id: locationId,
+    })
     .then(locationResponse =>
       dispatch({
         type: GET_LOCATION_RESPONSE,
         payload: locationResponse.data,
       }))
     .catch(e => console.error('error', e));
+};
+
+export const getTaxonomy = () => (dispatch) => {
+  api
+    .getTaxonomy()
+    .then(({ data }) =>
+      dispatch({
+        type: GET_TAXONOMY_RESPONSE,
+        payload: data,
+      }))
+    .catch(e => console.error('error', e));
+};
+
+export const selectCategories = (locationId, params) => (dispatch) => {
+  dispatch({
+    type: SELECT_CATEGORIES_ACTION,
+    payload: params,
+  });
 };
 
 export const updateLocation = (locationId, params) => (dispatch) => {
@@ -34,10 +50,11 @@ export const updateLocation = (locationId, params) => (dispatch) => {
       params,
     },
   });
-  updateLoc({
-    id: locationId,
-    params,
-  })
+  api
+    .updateLocation({
+      id: locationId,
+      params,
+    })
     .then((locationData) => {
       // do nothing, because save succeeded
     })
@@ -63,10 +80,11 @@ export const updatePhone = (locationId, phoneId, params) => (dispatch) => {
       params,
     },
   });
-  updatePh({
-    id: phoneId,
-    params,
-  })
+  api
+    .updatePhone({
+      id: phoneId,
+      params,
+    })
     .then((phoneData) => {
       // do nothing, because save succeeded
     })
@@ -91,10 +109,11 @@ export const createPhone = (locationId, phoneId, params) => (dispatch) => {
       params,
     },
   });
-  createPh({
-    id: locationId,
-    params,
-  })
+  api
+    .createPhone({
+      id: locationId,
+      params,
+    })
     .then((response) => {
       dispatch({
         type: CREATE_PHONE_SUCCESS,
